+++ includes/Migrations/AnyCommentMigration_0_0_64.php
@@ -9,7 +9,7 @@ class AnyCommentMigration_0_0_64 extends AnyCommentMigration {
 	/**
 	 * {@inheritdoc}
 	 */
-	public function isApplied() {
+	public function is_applied() {
 		return version_compare( get_option( 'anycomment_migration' ), $this->version, '>=' );
 	}