+++ includes/Migrations/AnyCommentMigration_0_0_66.php
@@ -9,7 +9,7 @@ class AnyCommentMigration_0_0_66 extends AnyCommentMigration {
 	/**
 	 * {@inheritdoc}
 	 */
-	public function isApplied() {
+	public function is_applied() {
 		global $wpdb;
 
 		$table = $wpdb->prefix . $this->table;